Sprint 6 - Commit 29/03/2022 @ 02:45
Combat system fully implemented. Honestly, I am not proud of how I implemented the link between weapons and their corresponding skills, but honestly, I wanted to just implement it to just cross it off the list of stuff to do for this week. After my last feature is added to the game (quest system), I will be conducting plenty of code refactoring in the run-up to the showcase day.